requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0

# Configuration and serialization
pyyaml>=6.0
pydantic>=1.8.0

# API framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# Export formats
pyarrow>=5.0.0  # for parquet
tables>=3.6.0   # for HDF5
h5py>=3.1.0

# Optional database support
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.7.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.0.0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910

# Production server
gunicorn>=20.1.0
